feat(sdk): implement support for gear vr
This is the PR #1033 updated to the latest VRTK version,
modified to be compliant with the latest changes to SDK management.

Refactor and update GearVR/SDK files (including latest changes to
`SDK_BaseBoundaries`).

Update documentation (`GearVR_README.md`).

Tested with GearVR:
* object touching and grabbing (005_Controller_BasicObjectGrabbing)
* interaction with Unity UI (034_Controls_InteractingWithUnityUI)
* teleport (044_CameraRig_RestrictedTeleportZones)

The examples were tested after changing `[CameraRig]` and `[VRTK]` in
order to support GearVR.