fix(Interaction): prevent error when exiting after haptics
There was an out of sync error when the scene was exited after
haptics had been started because the dictionary was trying to
be looped with a foreach but lost order during the shut down.

This has been fixed by simply just calling `StopAllCoroutines`
instead of stopping each one individually.