docs(Deprecation): remove internal deprecation warnings
The deprecated symbols correctly use the C# Obsolete Attribute, but
internal usage of deprecated things in VRTK meant warning messages were
loggeed in IDEs and Unity's console all the time.
Since those internal usages are of no interest to users of VRTK this
change tells the IDE and Unity to ignore those.