chore(Structure): ensure vrtk components can be injected into scripts
Rather than have VRTK scripts simply assume another VRTK script is on
the same GameObject (or on an ancestor or descendant GameObject) it is
now possible to inject the dependent script via a public inspector
property.

This makes it easier specify which component the script should be using
rather than it always just getting the first component it finds on the
checked GameObject.