fix(ControllerReference): prevent getting reference in OnEnable
Because SteamVR disables the controllers at start up, it means that if
the ControllerReference is attempted to be retrieved in an OnEnable
method, then the controller reference will be invalid therefore
causing issues with the rest of the script. The better solution is
to use a Getter that attempts to get the reference a fresh every
time it is requested, as the lower level calls are all cached then
this won't cause any slow down issues and will ensure the controller
is available.