fix(SDK): ensure SteamVR boundary mesh is generated correctly
If the SteamVR PlayArea `Draw In Game` option is unchecked then the
play area mesh is not generated which means anything relying on the
vertices from the mesh will not work (such as the play area cursor).

This fix ensures the `BuildMesh` method is called in the boundary
init function so it is always called at least once at runtime.

The second issue came from when the play area size was set to a
specific size other than calibrated. It seems that SteamVR puts
the vertices into a different order when using non-calibrated sizes
as can be seen in this issue on the SteamVR Github page:

https://github.com/ValveSoftware/steamvr_unity_plugin/issues/25

This has been fixed within VRTK by flipping the vertices accordingly
so they are in the correct position. In the future, if Valve fix this
within SteamVR then this can be wrapped in a script define so it only
affects older versions of SteamVR.