fix(FPSViewer): ensure correct SDK camera is set as canvas camera
The FPS Viewer uses the `Screen Space - Camera` render mode for the
canvas, but would previously require the camera to be specified in
the editor. However, now with the hot SDK switching it may not be
possible to set the camera so the FPS Viewer script now looks up
the relevant camera and sets it up automatically.