feat(Pointer): add ability to hide play area cursor on invalid location
A new option has been added to the Play Area Cursor script that will
hide the play area cursor if the cursor location is invalid (e.g. it
does not match the policy list or a play area collision has occurred).

This is useful for only showing the play area cursor when the user
is able to move to the new destination.