fix(Tooltips): make retry init variables public
The initialisation retry attempts of the Controller Tooltips can now
be customised via the inspector parameters rather than being hidden
variables.

This is useful if they need further tweaking, or no auto initialisation
is required and then they can be disabled.

Sometimes controllers take longer to initialise so it's now possible
to increase the retry attempts to ensure initialisation of the
tooltips still works.