feat(SnapDropZone): add getter method for drop zone object
Two new methods have been added to get the current hovered objects
and the current snapped object from within a snap drop zone.

Previously, the only way of knowing what was in a snap drop zone was
to register for the event when an item was hovered or snapped into
the zone.

Now these two new methods provide a way of directly accessing the
info on the snap drop zone without needing to use the event.