feat(Presence): option to ignore game object collisions to body physics
The Body Physics script now has a new `Ignore Collisions With`
parameter that allows for a list of GameObjects to be provided that
will have their colliders excluded from colliding with the Body Physics
generated colliders.

This is useful if using the Hip Tracking script and the hip tracked
object collider is causing the Body Physics collider to be pushed.