fix(Pointer): reset destination hit info on activate
The `destinationHit` variable was not being reset after the pointer
was disabled, meaning if the pointer was deactivated over a valid
object and then reactivated over that same object then the
`PointerEnter` method would not call the `OnDestinationMarkerEnter`
event.

The fix ensures the `destinationHit` is cleared whenever the
pointer is toggled on so it has a fresh state.