feat(Simulator): add sprint and new distance grab mode
Add sprint modifier key + relevant settings for it.
Adjusted near clip plane from 30cm to 5cm.
Add new distance grab mechanic, when in "mouse look"
mode its possible to press a modifier key (default left shift)
to get a crosshair, when aiming that crosshair on a
object with InteractibleObject on it, that object will
get picked up depending on if left or right hand key
was pressed.