feat(Locomotion): add option to ignore snap to floor on teleporter
The height adjust teleporter would always snap the play area to the
nearest floor, however it may be required to not snap to the nearest
floor and to snap to where ever the Y destination of the destination
marker is given.

This feature adds a new checkbox on the height adjust teleporter
to allow it to ignore the snapping to the nearest floor.