fix(SDKManager): honour script enable states
The VRTK scripts that register with the SDK Manager were always
being set to enabled even if the script was set to disabled in the
editor.

This fix stores the state of the registered script and then sets it
back to that original state when the SDK Manager runs through the
behaviours registered with it. This means the enabled state of the
scripts are now remembered and set appropriately.