fix(Locomotion): ensure snap to nearest floor is called with climbing
The Climbing script now forces the Snap To Nearest Floor code within
the Body Physics script when a climbable object is released. This
ensures that when the user releases a climbing object whilst they
are clipping the top of the object then they will still be teleported
to the top of the valid surface rather than having the collision
resolved and being thrown back down to the ground.