fix(Locomotion): ensure rewind on climb collision resets correctly
The Position Rewind script is now utilised better by the Player Climb
script to reset the position of the play area to a valid safe location
if the user ungrabs a climbable object whilst clipping inside a
collider.

The Position Rewind script has been updated so the valid last safe
position can be manually set and the Rewind Position method can
be manually called, which they are by the Player Climb script so
it can resolve the position issues before the Unity collision
resolution attempts to sort out any collisions.