fix(Interactions): pass relevant VRTK_InteractX to interactable object
The Interactable Object script would previously accept a GameObject
when the Touching/Grabbing/Using methods were called. However, this
was ambiguous as to what the object should represent. The object
that is performing the action should be the one with the relevant
VRTK_InteractX script on like VRTK_InteractTouch, VRTK_InteractGrab
and VRTK_InteractUse.

Because of this, it makes more sense to actually pass the reference
to the script as the GameObject can be obtained from the script
with less overhead than the script being obtained from the GameObject
via the `GetComponent` method call.

The existing methods that accept a GameObject still exist but have
been deprecated in favour of the new methods that take the specific
InteractX script.