fix(SDK): ensure simulator returns correct transform for headset
The Simulator now has a new Neck container for the camera but the
SDK for the simulator had not been updated to return the Camera
object from within the new Neck parent. This has now been updated
so the Camera is looked for within the Neck object.

The SDK Setup prefab has also been updated to manually populate
all of the Simulator settings as to prevent a crash where the SDK
Manager cannot look up the Simulator elements on first scene run.