fix(Pointers): ensure object interactor is set by global scale
The pointer object interactor that becomes an extension of the
controller but tracks the scale of the actual pointer cursor.

This would cause an issue when the camera rig was scaled, the
object interactor would change size along with the camera rig
and no longer match the scale of the pointer cursor.

This fix uses the global scale of the pointer cursor to set the
global scale of the object interactor so the scales continue
to match even when the camera rig is scaled.