feat(Presence): provide collision sweep detection for body physics
The Body Physics script now has a new method `SweepCollision` that
can be used to determine if the body physics body collider will
collide with an object if it continues to move in a given direction
for a given distance.

This new method is now used in some of the Locomotion methods that
move the play area via `transform.position` and can now check
if they are provided with a Body Physics script if when they move
forward, will the new position cause collision tunnelling and if
it does then the movement can be denied meaning tunnelling is not
possible.