feat(Utilities): ability to modify transform orientation based on sdk
The new SDK Transform Modify script can be used to change the
orientation of a GameObject transform determined by the current
loaded SDK.

This is useful if transform orientations differ between SDK setups
or different controllers within the SDK (and cross other SDKs).

The main use for this is to ensure that snap handles on interactable
objects can be positioned accordingly for each SDK controller, for
example the snap handle on the SteamVR Vive wand is at a different
orientation to that of the Oculus Touch controller.