feat(Pointer): allow valid/invalid color to change direction indicator
The Pointer Direction Indicator now allows for the pointer valid and
invalid colour to change the colour of the direction indicator.

Another option to also hide the direction indicator on an invalid
location has been added to work in a similar way to how the
play area cursor script works.