chore(CONTRIBUTING.md): provide instructions for new branch structure
The new branch structure for contributing code is now documented.
The new structure requires all bug fixes to have the pull request
raised on the `master` branch but any new features require the
pull request to be raised on the next relevant feature branch.

This means that patch releases are much easier to do as the master
branch is only used for fixing bugs and therefore patch releases
can be done off the master branch. Then the master branch will be
rebased into the next release branch which will also contain new
features.

When the next release branch is ready for release it will be
rebased back into the master branch and a release created.
Then the current next release branch will be deleted and a new
next release branch will be created.