fix(Presence): reduce GetComponent calls in body physics update
The BodyPhysics script was doing a `GetComponent` call every
`FixedUpdate` when checking to see if the snap to floor was valid
by checking to see if the object being held was underneath the
headset.

The fix attempts to cache the check the first time around to reduce
any further unnecessary calls to `GetComponent` if the held object
has not changed.