fix(Locomotion): replace `timeSinceLevelLoad` to prevent stuck timer
The SnapRotate and WarpObject control actions both used
`Time.timeSinceLevelLoad` to manage the action delay, however this
would cause an issue when switching scene as the time would no longer
be correct when calculating the delay.

The fix is simply to use `Time.time` to determine when the delay has
subsided.