fix(Logger): prevent custom logger being destroyed on awake
The Logger script was destroying any custom logger game object on
awake due to a change in the Logger where it would initialise at
edit time and no longer at runtime. This meant that the instance
was already set and therefore in the `Awake` method would just be
destroyed.

This fix just captures the instance of the script on the GameObject
and sets it to the active instance. If no Logger script is
applied to a scene GameObject then it will just use the hidden
one that is created at edit time.