fix(Pointer): ensure material change is done correctly
The `ChangeMaterial` method in the BasePointerRenderer script was
only being called if the current colour was different to the previous
colour, however it should only be called if the current colour is
not `Color.clear` otherwise it can be set incorrectly.

This fix ensures the pointer colours are set correctly.