feat(DesktopCamera): render desktop-only camera while in VR
There's a need to show something different on the desktop window
than what's shown in the VR headset. The potential use cases range
from offering nicer spectator views to showing completely different
views for asymmetrical games.
In combination with VRTK's existing ObjectFollow scripts this allows
for trailer recordings in which the VR headset's movement is
smoothed.

Having a separated desktop-only camera allows for:
- additional overlays or effects
- render other layers in addition/only
- changing the FOV
- ...

This change adds the new `VRTK_DesktopCamera` script and a
corresponding prefab that shows how to use the optional support for
rendering the headset's view in addition to the desktop-only view.