fix(SDKManager): deprecate persist on load setting
When using the persist on load setting on the SDK Manager it is
moved to the `DontDestroyOnLoad` scene at runtime by Unity. That
scene is inaccessible which means there's no way to find objects in
that scene. Because of the issues related to looking up objects in
the inaccessible scene the persist on load setting on the SDK
Manager is now deprecated.

It's recommended to use additive scene loading instead. To reduce
the likelihood of these issues breaking the project for users
already using the setting this fix also makes sure to only populate
the SDK Setup object references when this setting is *not* checked.