fix(SDKManager): handle null Behaviour
The SDK Manager offers to remember any Behaviour to automatically
toggle it off and on based on whether an SDK Setup is loaded to
remove the need to listen to the SDK Setup Loaded event manually.
When users forget to call the remove method the SDK Manager will
throw a null reference exception which doesn't offer the user enough
info.

This fix handles that case and logs an error in addition to removing
the bad Behaviours.