fix(Internal): emit tracked controller index change events
There was an issue with the code that checks to see if a controller
index has changed which was preventing the index changed event
from being emitted.

This issue has now been fixed by ensuring the index is set to the
newly found index.

It's also not appropriate to only emit events when their index is
less than the `uint.MaxValue` because a controller event can still
be valid even if the index has become the max value.