fix(Presence): calculate collider height based on global positions
There was an issue with the Body Physics body collider being sized
incorrectly in the simulator because the calculation was relying
on local positions and the simulator has a more complex set up
with the new neck object.

This fix uses global positions and is more generic to provide a
better future proof solution.

Thanks to @virror for the fix.