fix(Interaction): ensure interactable object rigidbody is set
There was an issue with the Interactable Object attempting to set
the rigidbody variable. As it was set in the `Awake` method and
only attempted to be set once, there could be times when the rigidbody
was checked for existence and it didn't exist.

This was present in the Snap Drop Zone if an Interactable Object was
snapped to a drop zone on scene start and the Interactable Object's
rigidbody was set to kinematic. The part of the Interactable Object
that saves the state of the rigidbody wasn't being run because
the rigidbody hadn't been set at that time, even though it was called
in `Awake`.

The fix now implements a new method that checks to see if the rigidbody
has been set and if not attempts to set it before returning if it
is valid or not.