feat(Avatar): add basic avatar hand prefab with controller script
A new set of avatar hand models has been added with a controller script
for animating movement of the fingers on the hand models.

It's also possible to provide custom finger axis values based on the
current interaction type (e.g. touch, grab, use).

The CustomControllerModel example scene has been updated to use the new
avatar hands as an example of how to use them.

Big thanks to @jimthegrim in the VRTK slack channel for all of the
hard work on the hand models, animations and figuring out the
layers and weights work in Unity to make these hands possible. <3