fix(CustomRaycast): set default raycast layer mask to correct setting
There was an issue where the CustomRaycast default raycast layer mask
was set to a blank LayerMask which meant all raycasts that were not
custom were not ignoring the default `Ignore Raycasts` layer.

This caused a number of issues such as pointers colliding with the
pointer cursor collider as it was set to the ignore raycast layer
but the raycast from the pointer wasn't being set to ignore
raycasts.

The fix is to ensure the default LayerMask is set to the
`Physics.IgnoreRaycastLayer`.