fix(Interaction): prevent double unuse event
The Interact Use script was emitting two unuse events because the
Interact Use script calls StopUsing on the Interactable Object and
the Interactable Object script calls ForceResetUsing on the
Interact Use script, which would then in turn emit the unuse event
before the original StopUsing routine had finished. This would
result in two unuse events being emitted.

The fix for this is to add another check into the Interactable Object
StopUsing method that can be used to bypass the ForceResetUsing call
as it's not needed to reset the Interact Use object if the StopUsing
on the Interact Use has been called as it will already reset the state
anyway.