fix(Locomotion): get correct body physics collider in object control
The ObjectControlAction scripts were not always getting the correct
BodyPhysics collider if no BodyPhysics was available as they were
just looking for the child CapsuleCollider and this could exist
in any of the CameraRig children such as on the Controller Objects.

This fix ensures it uses the BodyPhysics script and get the body
collider to use in the ObjectControlAction scripts.