fix(Controls): prevent button anchor changing on enable
The button anchor was continually changing based on the initial
constant force that had been added to the button which meant
every time a button was disabled and re-enabled the position of
the button would move slightly each time.

This fix allows the anchor to be auto configured the first time
but then after that it prevents it from be auto set so it stays
at the correct position.