fix(SDK): prevent null exception if no controller model found
The SteamVR Controller SDK could potentially throw an exception
if no controller model was found because the `Transform.Find` method
could potentially find `null` but the `gameObject` getter was being
chained on the end of the `Find` call.

The fix is to split the line into two and do a null check before
attempting to get the gameObject.