fix(Pointer): child straight pointer renderer to smoothing object
There was an issue with the Straight Pointer Renderer with GearVR
and the Simulator where it would lag behind the controller due to
how the position of the pointer tracer was being updated to always
match the smoothing object position, which was in turn being matched
to the actual controller position with TransformFollow.

The solution is simply childing the pointer container to the
smoothing Game Object so it doesn't need the secondary update which
causes the lag.