fix(SnapDropZone): ensure highlight always active option is honoured
There was an issue withe the Highlight Always Active option where it
could be turned on at scene run, but then if it was turned off during
play it would not turn off the highlighter even though the state of
the option would dictate otherwise.

The fix is to ensure the `isHighlighted` flag is always correctly
set and an additional check is done to make sure to turn the
highlighter off if it shouldn't be on due to this option being false.