refactor(SDK): replace oculus angular velocity calculations
The Oculus SDK now uses the Velocity Estimator to calculate the
angular velocity of the controllers and the headset.

Even though the Oculus SDK has a `GetLocalControllerAngularVelocity`
method on `OVRInput` this seems to provide the rotations in an
incorrect direction, plus the documentation still says this is
supposed to return a Quaternion rather than a Vector3 (and it actually
returns a Vector3). Because of these omissions, it's not trustworthy
enough to use the build in Oculus SDK angular velocity method at
this time.