refactor(Structure): change documentation style for enums
The enum documentation has now been changed to the correct way of
documenting enum values for C# by using the `<summary>` tag above
each of the enum values.

Additional unsupported tags have been removed (for now) from the code
docs so the documentation generates correctly.

It would be appropriate to add this sort of inline tagging only when
the doc generator supports it and it is used universally throughout
all of the classes.