fix(Interaction): handle VRTK_SnapDropZone colliders correctly
The VRTK_SnapDropZone code was assuming that GameObjects with the
VRTK_InteractableObject component would also contain a collider.

It would be valuable to allow said GameObjects to have the relevant
collider attached to a child object instead of the parent object. The
VRTK_SnapDropZone code had no problems with the practice of storing
the object mesh filter and renderer on a child object, but it would
be valuable  to cease requiring that the collider be on the parent
object alongside the VRTK_InteractableObject component. If a developer
needs to tweak the location or scale of the child object (eg. in order
to correct the center point or the scale of a crappy Unity Store
asset), the collider will need to be attached to the child object along
with the visuals so that the renderer and collider bounds match.