feat(SDK): add actual controller ready events
The Controller SDK now emits an event when the actual Left and Right
controller GameObjects become available and active in the scene.

As soon as an actual controller GameObject is available, it should
be possible to get the controller type without the need for polling
via coroutines.

The check to see if the controller GameObject is available is done
via a polling coroutine in the SDK Manager, which checks every second
to see if the actual GameObject is active in the scene and if it is
then it emits the event. This event is also emitted every time the
SDK loaded setup changes, so it will notify the controller SDK when
a new SDK has loaded and the SDK controllers are available.

The SDK Object State script and the SDK Transform Modify script now
both use this new event rather than polling the state of the controller
type, which was inefficient and didn't work very well as it would
eventually stop trying after a number of attempts.

This new poll never stops trying, but is limited to 1 try every
second which should reduce overhead and is just a simple boolean check
on whether the actual controller GameObject is active.