feat(SnapDropZone): use interaction events for snapping
The SnapDropZone has been refactored to now utilise the
interactable object over a simple GameObject. This also
enables the ability to listen for the ungrab event on the
interactable object and then determine the snap on that event
rather than constantly polling in the OnTriggerStay method for
the snap event.

This also means that the default snapped object has been deprecated
and an interactable object is provided as the default.

Only interactable objects can be snapped so this makes more sense.