fix(Interaction): prevent null exception when getting touch colliders
The Interact Touch script could potentially return a null exception
error when calling the `ControllerColliders` method because if the
condition was false it would still try and perform a `GetComponent`
on a null object.

It was also not necessary to do the GetComponents or
GetComponentsInChildren check as the latter will also get components
in the current object as well as the children.