feat(Interaction): extract interactable object highlight to new script
The Interactable Object script has now had the `Touch Highlight`
parameter deprecated and a new script called Interact Object
Highlighter that deals with highlighting the Interactable Object based
on any of the interaction types.

This allows for highlighting on near touch, touch, grab and use without
any of that logic needing to sit within the Interactable Object. The
ties to the Interactable Object highlighter in other scripts like
Interact Touch and Interact Grab have also been removed as they are
no longer required to manage the highlighter as the new script uses
the Interactable Object interaction events to determine when to
manage the highlighter.

The InteractableObject `ToggleHighlight` method has also been
deprecated in favour of two new methods `Highlight` and `Unhighlight`
that make it explicitly clear of how to use the methods.

The Object Setup editor script has also been tidied up and now
generates the required Interac Object Highlight script if required.