fix(Interaction): ensure valid grab attempt is called
The Interact Grab script requires the `IsValidGrabAttempt` method
to be called to set up state for future method calls.

This was removed during a cleanup that said the line was not being
used, but the variable it was being assigned to wasn't being used,
however, the method call was still required.