fix(SnapDropZone): prevent unwanted unsnap on leaving trigger collider
The Snap Drop Zone will automatically unsnap an object if it leaves the
trigger collider of the snap drop zone. However, this isn't always the
correct result. If the snap drop zone is using a joint snap and the
joint is set to infinity, then it would never be expected to break the
joint. Also, if the interactable object can only be dropped in snap
drop zones then forcing it out of the snap drop zone when it leaves
the trigger collider is also incorrect.