fix(SnapDropZone): allow grab remove from unsnappable joints
A previous commit made it so it was not possible to unsnap objects
from Snap Drop Zones when they left the trigger collider if they were
considered objects that should not be able to be unsnapped.

However, this caused an issue that it was not possible to grab objects
out of snap drop zones either without releasing them and having them
snap straight back into the snap drop zone.

This issue ensures any grabbed object can be removed from a snap drop
zone.