feat(Pointers): auto create pointer renderer if not set
The Pointer script will now attempt to find a Pointer Renderer
on the same GameObject or any children if the Pointer Renderer
parameter is not set. If it cannot find one then it will auto
generate a Straight Pointer Renderer.

This means if a Pointer is not configured correctly, then at least
some functionality will be available preventing the assumption
that things are broken.