fix(Structure): ensure script references are correctly set
Previously, many of the script references required scripts to be placed
on the same GameObject as the existing script. Things like controller
related scripts should be able to be placed on child GameObjects on
the script alias objects.

Any generic play area script should not have to be on the same
GameObject as any other script so doing a `FindObjectOfType` is a
better way of finding the script. Even though this is slower it is
only done in `OnEnable` so shouldn't be that slow.

If it is slow then the ObjectCache can be used in the future.