feat(Interaction): allow different rotation actions for rotate grab
The Rotate Transform Grab Mechanic now has a new rotation action option
which allows the object to be rotated based on not only the angle in
difference between the origin of the object and the controller current
grab position but can also be based on the rotation of the controller
through one of it's axes (roll, pitch, yaw).

This can help with simulating other mechanics such as turning a screw
by swiveling the controller in the hand instead of grabbing and pulling
the object in the direction to move it.