fix(UIPointer): exit events per element
The VRInputModule script would not fire `UIPointerElementExit` if the
pointer landed on a new element instead of nothing. This fix makes
sure to fire the exit event on any existing `hoveringElement` before
firing enter on the new element. The other case for the exit event
firing is left as is as it handles the situation where you move off
all elements and there are no results or target.