fix(Utilities): ensure controller ready event happens on sdk switch
The Controller Ready event wasn't being emitted correctly due to the
SDK switching happening when the scene first starts. This meant that
the SDK Transform Modify and SDK Object State script were not working
on SDK switch correctly.

The solution is to ensure the Controller Ready event is always
unregistered and re-registered when then SDK setup changes.

This now happens in a new base script called SDK Controller Ready and
the relevant scripts that need to deal with the controller being ready
can now just extend that script and have access to the controller ready
event even when the SDK setup switches.

The Device Finder has also had a few more methods that make it easy
to find the Controller Reference of the left or right hand.