fix(Interaction): prevent object highlighter always being created
The Interactable Object now relies on the Interact Object Highlighter
to handle highlighting and previously if a highlight colour wasn't
provided then the base Material Colour Swap highlighter wasn't
automatically added to the Interactable Object GameObject.

Since the Interact Object Highlighter was added, the base Highlighter
was always being added to the Interactable Object even if no
highlighting was occurring.

The fix is to only initialise the highlighter if a highlight colour
is available (i.e. not set to Color.clear).