fix(TrackObjectGrabAttach): dropping objects early
maxDistanceDelta being a constant caused the magnitudes of the
calculated positional and angular velocities to not be limited properly.
This resulted in the held objects getting dropped prematurely without
properly detaching them: The objects were attached in theory but not
following the controller as intended.

The solution is to change the constant to the predefined velocity limit,
causing the magnitude to limit properly to fit the later if-statements.